# swagger.yaml
swagger: '2.0'
info:
  description: 'Assisted installation'
  version: 1.0.0
  title: AssistedInstall
host: api.openshift.com
basePath: /api/assisted-install/v1
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  agentAuth:
    type: apiKey
    in: header
    name: X-Secret-Key
  userAuth:
    type: apiKey
    in: header
    name: Authorization
security:
  - userAuth: [admin, user]
paths:
...
  /assisted-service-iso:
    post:
      tags:
        - assisted-service-iso
      summary: Creates ISO for the user and uploads to S3.
      operationId: CreateISOAndUploadToS3
      parameters:
        - in: body
          name: assisted-service-iso-create-params
          required: true
          schema:
            $ref: '#/definitions/assisted-service-iso-create-params'
      responses:
        200:
          description: Success.
          schema:
            $ref: '#/definitions/presigned'
        400:
          description: Error.
          schema:
            $ref: '#/definitions/error'
        401:
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        403:
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        500:
          description: Error.
          schema:
            $ref: '#/definitions/error'
  /assisted-service-iso/data:
    get:
      tags:
        - assisted-service-iso
      summary: Downloads the Assisted Service ISO.
      security:
        - userAuth: [admin, read-only-admin, user]
      operationId: DownloadISO
      produces:
        # application/vnd.efi.iso is not supported
        - application/octet-stream
      responses:
        200:
          description: Success.
          schema:
            type: string
            format: binary
        401:
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        403:
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        404:
          description: Error.
          schema:
            $ref: '#/definitions/error'
        500:
          description: Error.
          schema:
            $ref: '#/definitions/error'
  /assisted-service-iso/presigned:
    get:
      tags:
        - assisted-service-iso
      security:
        - userAuth: [admin, read-only-admin, user]
      summary: Retrieves a pre-signed S3 URL for downloading assisted-service ISO.
      operationId: GetPresignedForAssistedServiceISO
      responses:
        200:
          description: Success.
          schema:
            $ref: '#/definitions/presigned'
        401:
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        403:
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        404:
          description: Error.
          schema:
            $ref: '#/definitions/error'
        500:
          description: Error.
          schema:
            $ref: '#/definitions/error'
definitions:
...
  assisted-service-iso-create-params:
    type: object
    properties:
      ssh_public_key:
        type: string
        description: SSH public key for debugging the installation.
      pull_secret:
        type: string
        description: The pull secret obtained from Red Hat OpenShift Cluster Manager at cloud.redhat.com/openshift/install/pull-secret.
...
  error:
    type: object
    required:
      - kind
      - id
      - href
      - code
      - reason
    properties:
      kind:
        type: string
        enum: ['Error']
        description: Indicates the type of this object. Will always be 'Error'.
      id:
        type: integer
        format: int32
        description: Numeric identifier of the error.
        minimum: 400
        maximum: 504
      href:
        type: string
        description: Self link.
      code:
        type: string
        description: Globally unique code of the error, composed of the unique identifier of the API and the numeric identifier of the error. For example, if the numeric identifier of the error is 93 and the identifier of the API is assisted_install then the code will be ASSISTED-INSTALL-93.
      reason:
        type: string
        description: Human-readable description of the error.
  infra_error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        description: Numeric identifier of the error.
        minimum: 401
        maximum: 403
      message:
        type: string
        description: Human-readable description of the error.
...
  presigned:
    type: object
    required:
      - url
    properties:
      url:
        type: string
...